add a script to generate the configmap for metal3
The metal3 application needs some settings from the installer that are
not being passed through to it directly, yet. This script generates a
manifest for a ConfigMap to set those values. It should be used to
create the manifest as part of running the installer, because the
installation will not complete without the ConfigMap in place.

Addresses #30