Start adding 03_create_cluster.sh.
This is the beginning of a create_cluster script.  It knows how to
extract the baremetal installer from an OpenShift release image and
run it.

There are still some missing pieces, most notably some help in
generating a complete install-config.yaml with the details of the bare
metal hosts.  Known todo items are listed with links to their
corresponding issues.