Make it easier to drop the release override.
This is based on feedback from #35.

The release override can be dropped once the oc changes merge that
allow us to extract the installer with the release image set
appropriately.  Move the release override var to one spot where it'll
easily be removed when the time comes.

Also drop "create manifests", as we don't use it yet, and it's unclear
if we'll keep it at all.